requirements.txt
requests
pandas
pyarrow
psycopg2-binary
python-dotenv
